requirements.txt
# Core ML/AI Libraries
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
huggingface-hub>=0.21.0

# Agent Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# UI/Demo
gradio>=4.0.0
streamlit>=1.30.0

# API Framework (optional)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Logging and Monitoring
loguru>=0.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter for Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
tenacity>=8.2.0

# Optional: Model Quantization (for edge deployment)
bitsandbytes>=0.41.0
optimum>=1.16.0
